refactor: model 파일내 파일명 model 제거
