+++ frontend/src/models/weather.ts
